# openapi.yaml
openapi: 3.0.3
info:
  title: Sports Companion API
  version: 1.0.0
servers:
  - url: https://api.sportification.app/v1
paths:
...
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Get user
      responses:
        '200':
          description: User data
    patch:
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated
  /matches:
    get:
      summary: List matches
      responses:
        '200':
          description: Successful response
    post:
      summary: Create match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Match'
      responses:
        '201':
          description: Match created
  /matches/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Get match
      responses:
        '200':
          description: Match data
    patch:
      summary: Update match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Match'
      responses:
        '200':
          description: Match updated
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        profile: { $ref: '#/components/schemas/Profile' }
        preferences: { type: object }
        stats: { type: object }
        friends:
          type: array
          items: { type: string }
        followers:
          type: array
          items: { type: string }
        achievements:
          type: array
          items: { type: string }
    Match:
      type: object
      properties:
        id: { type: string }
        type:
          type: string
          enum: [public, private]
        status:
          type: string
          enum: [upcoming, ongoing, completed, expired]
        participants:
          type: array
          items: { type: string }
        schedule: { type: object }
        venue: { $ref: '#/components/schemas/Venue' }
        rules: { type: object }
        chat: { type: string }
...
    Profile:
      type: object
      properties:
        id: { type: string }
        avatar: { type: string }
        bio: { type: string }
        achievements:
          type: array
          items: { type: string }
        qrCode: { type: string }
...
    Venue:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        location: { type: object }
        surfaceType: { type: string }
        capacity: { type: integer }
...